Move some replace.go functions back to lib/services.
